Move rest_api into mod with files for parse and req/resp
